agentic-rag-system/requirements.txt
# --- Embeddings, Re-ranking & Math ---
sentence-transformers>=2.5.0
rank-bm25>=0.2.2
numpy>=1.24.0

# --- Vector Database ---
qdrant-client>=1.8.0

# --- Ingestion & Parsing ---
unstructured[all-docs]>=0.12.0

# --- LLM Provider & Env Management ---
openai>=1.14.0
python-dotenv>=1.0.1

# --- Optional Frontend/Backend APIs ---
fastapi>=0.110.0
uvicorn>=0.29.0
gradio>=4.20.0